Do not overwrite Solr config if it exists.
git-svn-id: https://opensource.kantega.no/svn/openaksess/trunk@4893 fd808399-8219-4f14-9d4c-37719d9ec93d

+++ modules/solr-search/src/java/no/kantega/openaksess/search/solr/config/SolrConfigInitializer.java
@@ -77,8 +77,12 @@ public class SolrConfigInitializer {
     }
 
     private static void copyToConfigDir(ConfigPair configPair) throws IOException {
-        log.info("Copying {} to {}", configPair.resourcePath, configPair.targetFile);
-        copyAndCloseStreams(SolrConfigInitializer.class.getResourceAsStream(configPair.resourcePath), new FileOutputStream(configPair.targetFile));
+        if (configPair.targetFile.exists()) {
+            log.info(configPair.targetFile.getAbsolutePath() + " exists, will not overwrite");
+        } else {
+            log.info("Copying {} to {}", configPair.resourcePath, configPair.targetFile);
+            copyAndCloseStreams(SolrConfigInitializer.class.getResourceAsStream(configPair.resourcePath), new FileOutputStream(configPair.targetFile));
+        }
     }
 
     private static void copyAndCloseStreams(InputStream resourceAsStream, FileOutputStream output) throws IOException {
